requirements.txt
bcbio-gff==0.7.0
biopython==1.79
h5py==3.1.0
numpy==1.19.2
pandas==1.1.5
tqdm==4.64.1
torchmetrics==0.8.2